got first couple of subscribe tests using sinon spy, refactored out mount so I could re-use it
