+++ cypress/integration/inject.js
@@ -1,38 +1,11 @@
 import React from 'react'
 import { Inject, useInstance, withDependencies } from '../support/sut';
-import { ContainerContext, ContainerContextWithValue } from '../support';
+import { ContainerContext, ContainerContextWithValue, mount } from '../support';
 import { fixReactDOMScope } from '../support';
 import Sinon from 'sinon';
 
 // We only do identity checks, so doesn't need any properties
 class MyClass { }
-
-/**
- * This class solves 2 problems
- * 1. we need a way to find our dom element, so we give it an id
- * 2. the top component we render must have a display name (a constraint of https://github.com/bahmutov/cypress-react-unit-test)
- */
-class ReactUnitTestRoot extends React.Component {
-  render(){
-    return (
-      <div id="react-unit-test-root">
-        {this.props.children}
-      </div>
-    )
-  }
-}
-
-const mount = (children) => {
-  const App = (
-    <ReactUnitTestRoot>
-      {children}
-    </ReactUnitTestRoot>
-  );
-  cy.mount(App);
-
-  // return the dom element so that the consumer doesn't need to know what id we used
-  return cy.get('#react-unit-test-root');
-}
 
 describe('Injecting a dependency', () => {
